+++ lib/src/widgets/term.dart
@@ -78,7 +78,7 @@ class _TermUiState<T> extends State<TermUi<T>> {
       style: const TextStyle(fontSize: 14),
       decoration: InputDecoration(
         isDense: true,
-        contentPadding: EdgeInsets.symmetric(horizontal: 8.0, vertical: 10.0),
+        contentPadding: EdgeInsets.symmetric(horizontal: 8.0, vertical: 8.0),
         errorText: error,
       ),
       controller: controller,